--- hscontrol/protocol_common_utils.go
@@ -1,1 +1,0 @@
-package hscontrol
